requirements.txt
langchain
requests
langchain-openai
langchain-community
python-dotenv
streamlit
streamlit-chat
altair<5
langchain-core
langchain-text-splitters
pypdf
faiss-cpu
langchainhub
langsmith
langgraph
langchain-pinecone
requests
black
openai
google-search-results
python-dotenv
pinecone
beautifulsoup4
pinecone-client
firecrawl-py
